Fix startup issues by refactoring into startup
