+++ Samples/WindowsCopilotRuntime/cs-wpf-sparse/WCRforWPF/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:WCRforWPF"
-             StartupUri="MainWindow.xaml">
+             Startup="OnStartUp">
     <Application.Resources>
          
     </Application.Resources>
